data-producer/requirements.txt
boto3==1.34.51
botocore==1.34.51
PyYAML==6.0.1
